pyannotation/requirements.txt
django==4.2.11
djangorestframework==3.15.1
gunicorn==23.0.0